pyproject.toml
[project]
name = "bb-bugs"
version = "0.1.0"
description = "Fetch and analyze the BuzzerBeater bugs forum threads and store first posts."
readme = "README.md"
authors = [
  { name = "Ioannis Michaloliakos", email = "ioannis.michalol@ufl.edu" }
]
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12",
  "lxml>=5.1",
  "fastapi>=0.112.0",
  "uvicorn>=0.30.0",
  "python-dotenv>=1.0.0",
  "rich>=14.1.0",
  "requests>=2.31",
  "matplotlib>=3.10.8",
]

[project.scripts]
bb-bugs-fetch = "bb_bugs.cli:main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = []
